Test Version: use SPIN for multiple bus
This is a test version of this library, that uses my (KurtE) SPIN
library to remove some of the differences between SPI, SPI1, and SPI2 on
the new processor boards.  A lot of this comes from my ILI9341_t3n
version of the Pauls fast version.

In this version, I use SPIN for all T3.x boards.  SPin has in it code to
know which pins are valid for MOSI, MISO, SCK and I use that instead of
hard coding.

I added an optional SPIN object pointer to the constructor, so you can
directly say which SPI buss it is on (like my  ILI.._t3n), but an
addition, I added some code that if the pins passed in are not valid on
the main SPIN object and SPIN1 and SPIN2, exists, it updates to use
those objects instead.

Also like my ILI9341_t3n class, I default MOSI an SCK to pins 255 and
code checks if they are that, to not set the MOSI or SCK on the buss,
but just use the default (or current) ones.

Also code in place that allows the code to work if DC is on an hardware
CS pin for the proper buss, but the CS pin is not.  Sumotoy already had
this for SPI1 case as there is only 1 valid one unless you use the
SDCard pins.

Warning: This compiles, but I have not yet tested as I don't have the
hardware.  Should have a display hopefully within the week.